store size: treeshakeable optimistic channel + dead-export sweep
Moves the optimistic-only machinery (notifyOptimisticWrites,
consumeOverridesNext, optimisticView, applyTentative) out of the plain
store/reconcile modules into next/optimistic.ts behind an injection
table installed in createOptimisticStore's once-guard (no module-scope
side effects — the module shakes). Sound by construction: every call
site is fam?.opt-gated and optimistic families are only mintable by
createOptimisticStore. Dead-export sweep rides along. Trade-off ruled
(Ryan): plain store entry −481 gzip (13710 → 13229), full-import
bundle +181 (headline eaten); package remains −722 vs main. Perf
guards: dbmon tick next-wins 41/16, optimistic write-storm 0.94x,
suite + full monorepo green.
